Extract type definitions into types.go
- Created new types.go file with all core type definitions
- Moved AddrModeType, CPUVariant, Flags, Instruction, and State types
- Moved all associated constants and methods
- Moved FormatFlags helper function
- Removed duplicate definitions from cpu.go
- All tests pass successfully
- No breaking changes to public API
