+++ cpu.go
@@ -75,96 +75,6 @@ func (h *LoggingErrorHandler) HandleError(err *CPUError) error {
 	return nil // Continue execution
 }
 
-// AddrModeType represents the addressing mode of an instruction
-type AddrModeType uint8
-
-const (
-	AddrModeIMP AddrModeType = iota // Implied
-	AddrModeIMM                     // Immediate
-	AddrModeZP0                     // Zero Page
-	AddrModeZPX                     // Zero Page, X
-	AddrModeZPY                     // Zero Page, Y
-	AddrModeREL                     // Relative
-	AddrModeABS                     // Absolute
-	AddrModeABX                     // Absolute, X
-	AddrModeABY                     // Absolute, Y
-	AddrModeIND                     // Indirect
-	AddrModeIZX                     // Indexed Indirect
-	AddrModeIZY                     // Indirect Indexed
-)
-
-// String returns the addressing mode name for debugging
-func (a AddrModeType) String() string {
-	names := []string{
-		"IMP", "IMM", "ZP0", "ZPX", "ZPY", "REL",
-		"ABS", "ABX", "ABY", "IND", "IZX", "IZY",
-	}
-	if int(a) < len(names) {
-		return names[a]
-	}
-	return "UNKNOWN"
-}
-
-// CPUVariant represents different 6502 processor variants
-type CPUVariant int
-
-const (
-	// VariantNMOS6502 is the original NMOS 6502 (1975)
-	// Used in: Apple II, Commodore 64, Atari 2600/800, BBC Micro
-	// Features: All documented bugs, decimal mode supported
-	VariantNMOS6502 CPUVariant = iota
-
-	// VariantCMOS65C02 is the CMOS 65C02 (1982)
-	// Used in: Apple IIc, Apple IIe (enhanced), later systems
-	// Features: Bug fixes, additional instructions, lower power
-	VariantCMOS65C02
-
-	// VariantRicoh2A03 is the NES/Famicom CPU (1983)
-	// Used in: Nintendo Entertainment System, Famicom
-	// Features: No decimal mode, integrated APU, different timing
-	VariantRicoh2A03
-
-	// VariantRicoh2A07 is the PAL NES CPU
-	// Same as 2A03 but with PAL timing
-	VariantRicoh2A07
-)
-
-// String returns the variant name
-func (v CPUVariant) String() string {
-	names := []string{
-		"NMOS 6502",
-		"CMOS 65C02",
-		"Ricoh 2A03 (NTSC)",
-		"Ricoh 2A07 (PAL)",
-	}
-	if int(v) < len(names) {
-		return names[v]
-	}
-	return "Unknown"
-}
-
-// SupportsDecimalMode returns true if the variant supports decimal mode
-func (v CPUVariant) SupportsDecimalMode() bool {
-	switch v {
-	case VariantRicoh2A03, VariantRicoh2A07:
-		return false
-	default:
-		return true
-	}
-}
-
-// HasIndirectJMPBug returns true if the variant has the indirect JMP page boundary bug
-func (v CPUVariant) HasIndirectJMPBug() bool {
-	switch v {
-	case VariantNMOS6502, VariantRicoh2A03, VariantRicoh2A07:
-		return true
-	case VariantCMOS65C02:
-		return false
-	default:
-		return true
-	}
-}
-
 // Bus defines the interface for memory access.
 //
 // Implementations of this interface provide the CPU with access to
@@ -224,30 +134,6 @@ func DefaultConfig() CPUConfig {
 		InstructionCacheSize:   256,
 	}
 }
-
-// Flags represents the processor status register.
-//
-// The 6502 has 8 status flags that indicate the result of operations:
-//   - N (Negative): Set if result is negative (bit 7 = 1)
-//   - V (Overflow): Set if signed overflow occurred
-//   - U (Unused): Always set to 1
-//   - B (Break): Set when BRK instruction executed
-//   - D (Decimal): Enables BCD arithmetic mode
-//   - I (Interrupt Disable): When set, IRQ interrupts are ignored
-//   - Z (Zero): Set if result is zero
-//   - C (Carry): Set if unsigned overflow/borrow occurred
-type Flags uint8
-
-const (
-	C Flags = 1 << 0 // Carry Bit
-	Z Flags = 1 << 1 // Zero
-	I Flags = 1 << 2 // Disable Interrupts
-	D Flags = 1 << 3 // Decimal Mode (rarely used, often ignored in NES emu)
-	B Flags = 1 << 4 // Break Command
-	U Flags = 1 << 5 // Unused (always 1)
-	V Flags = 1 << 6 // Overflow
-	N Flags = 1 << 7 // Negative
-)
 
 // InstructionCacheEntry represents a cached instruction
 type InstructionCacheEntry struct {
@@ -376,18 +262,6 @@ type CPU struct {
 
 	// Performance optimization
 	instrCache *InstructionCache
-}
-
-// Instruction struct: Use method expressions for types
-type Instruction struct {
-	Name             string           // Mnemonic (e.g., "LDA")
-	Operate          func(*CPU) uint8 // Function to execute the instruction's logic (accepts *CPU)
-	AddrMode         func(*CPU) uint8 // Function to calculate the address and fetch data (accepts *CPU)
-	AddrModeType     AddrModeType     // Type of addressing mode
-	Cycles           uint8            // Base cycles for this instruction/mode
-	Length           uint8            // Length of the instruction in bytes
-	Illegal          bool             // Whether this is an official or unofficial/illegal opcode
-	PageCrossPenalty bool             // Whether to add +1 cycle on page boundary cross
 }
 
 // NewCPU creates a new 6502 CPU instance with default configuration.
@@ -1888,22 +1762,6 @@ func (c *CPU) EnableInstructionCache() {
 
 // --- State Inspection ---
 
-// State represents a snapshot of CPU state
-type State struct {
-	A               uint8
-	X               uint8
-	Y               uint8
-	SP              uint8
-	PC              uint16
-	P               Flags
-	Cycles          uint8
-	TotalCycles     uint64
-	Opcode          uint8
-	Instruction     string
-	InInterrupt     bool
-	InterruptVector uint16
-}
-
 // GetStateSnapshot returns a snapshot of the current CPU state
 func (c *CPU) GetStateSnapshot() State {
 	instrName := "???"
@@ -1925,15 +1783,6 @@ func (c *CPU) GetStateSnapshot() State {
 		InInterrupt:     c.inInterrupt,
 		InterruptVector: c.interruptVector,
 	}
-}
-
-// String returns a human-readable representation of the state
-func (s State) String() string {
-	return fmt.Sprintf(
-		"PC:%04X A:%02X X:%02X Y:%02X P:%02X[%s] SP:%02X CYC:%d (%s $%02X)",
-		s.PC, s.A, s.X, s.Y, uint8(s.P), FormatFlags(s.P),
-		s.SP, s.TotalCycles, s.Instruction, s.Opcode,
-	)
 }
 
 // --- Debug Helpers ---
@@ -2006,26 +1855,6 @@ func (c *CPU) GetState() string {
 
 	return fmt.Sprintf("PC:%04X A:%02X X:%02X Y:%02X P:%02X[%s] SP:%02X CYC:%d (%s $%02X)",
 		c.PC, c.A, c.X, c.Y, uint8(c.P), flagsStr, c.SP, c.totalCycles, instrName, c.opcode)
-}
-
-// FormatFlags - Helper to create the NVUBDIZC string
-func FormatFlags(p Flags) string {
-	flags := []struct {
-		flag Flags
-		char byte
-	}{
-		{N, 'N'}, {V, 'V'}, {U, 'U'}, {B, 'B'},
-		{D, 'D'}, {I, 'I'}, {Z, 'Z'}, {C, 'C'},
-	}
-	s := make([]byte, 8)
-	for i, f := range flags {
-		if (p & f.flag) != 0 {
-			s[i] = f.char
-		} else {
-			s[i] = '.'
-		}
-	}
-	return string(s)
 }
 
 // Disassemble - Disassemble instructions in memory range
